static_cast to remove warnings
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@3986 a380766d-ff14-0410-b294-a243070f3f08

+++ headers/lang/PyrSlot.h
@@ -227,10 +227,10 @@ extern PyrSymbol *s_super; // "super"
 inline int slotFloatVal(PyrSlot *slot, float *value)
 {
 	if (IsFloat(slot)) {
-		*value = slot->uf;
+		*value = static_cast<float>(slot->uf);
 		return errNone;
 	} else if (IsInt(slot)) {
-		 *value = slot->ui;
+		 *value = static_cast<float>(slot->ui);
 		return errNone;
 	}
 	return errWrongType;
